Remove semi-extraneous fpm --iteration from the package.
The iteration did need to be bumped to accommodate the dependency
change in the last commit.  I came up with 5 by looking at currently
published packages, which have a maximum iteration of 4.

However, our build scripts currently set the iteration to 5 already
for all Rails packages, and adding it here causes errors in the build
process because it comes between two package path specifications, for
the LICENSE file and the rest of the package.  So don't bother.
Refs #8905.