15426: Don't use 1 second timeout, explicitly sign out
The goal is to require the user to log in every time they arrive at
the SSO server.  We don't want long lived sessions.  This is essential
for the account linking feature, which requires logging in as user A
and then logging a second time as user B.

The reasoning behind the 1 second session timeout is that it seemed to
accomplish the goal of forcing login without breaking normal login
flow.  Except then the other day, login flow stopped working for
me. Confusingly, the login flow continues to work for other people,
and works on Chromium, but not Firefox.  I can't explain it.

The symptom is that the user becomes stuck in a login loop, because
when the user is redirected from the omniauth endpoint to the josh_id
endpoint the newly created session expires immediately.  Since the
user is no longer considered logged in, the user is redirected back to
Google.

The solution on this branch is to (a) increase the timeout to 1
minute, and (b) explicitly sign out after the browser is redirected
from the SSO server back to the API server to complete the login.